### 2025-03-15 #7900 @editorialbot
[PRE REVIEW]: ProbINet: Bridging Usability Gaps in Probabilistic Network Analysis
**Submitting author:** @diegoabt (<a href="http://orcid.org/0000-0003-2994-0138">Diego Baptista</a>)
**Repository:** https://github.com/MPI-IS/probinet
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
