### 2025-03-17 a248ee8
Fix ORCID number
Reduce the length (1493)

Keep reducing the length (1385)

Remove Mathematical background section here, and more (1193)

Remove even more (1079)

Remove even more (1054)

Add missing doi's
