+++ publication/paper.bib
@@ -6,6 +6,7 @@
   number={1},
   pages={15736},
   year={2020},
+  doi = {10.1038/s41598-020-72626-y},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -17,6 +18,7 @@
   number={2},
   pages={023209},
   year={2021},
+  doi = {10.1103/PhysRevResearch.3.023209},
   publisher={APS}
 }
 
@@ -28,6 +30,7 @@
   number={4},
   pages={cnac034},
   year={2022},
+  doi={10.1093/comnet/cnac034},
   publisher={Oxford University Press}
 }
 
@@ -39,6 +42,7 @@
   number={1},
   pages={015010},
   year={2022},
+  doi = {10.1088/2632-072X/ac52e6},
   publisher={IOP Publishing}
 }
 
@@ -50,6 +54,7 @@
   number={1},
   pages={122},
   year={2022},
+  doi={10.1186/s40537-022-00669-1},
   publisher={Springer}
 }
 
