+++ publication/paper.md
@@ -10,7 +10,7 @@ tags:
   - synthetic data generation
 authors:
   - name: Diego Baptista
-    orcid: 0000-0000-0000-0000
+    orcid: 0000-0003-2994-0138
     #equal-contrib: true
     affiliation: 1 
   - name: Martina Contisciani
@@ -37,21 +37,25 @@ bibliography: paper.bib
 
 **Prob**abilistic **I**nference on **Net**works (ProbINet) is a Python package that provides a 
 unified framework to perform probabilistic inference on networks, enabling researchers and practitioners 
-to analyze and model complex network data. The package integrates code implementations from several scientific publications, supporting tasks such as community detection, anomaly detection, and synthetic data generation using latent variable models. It is designed to simplify the use of cutting-edge techniques in network analysis by providing a cohesive and user-friendly interface. The package includes efficient implementations of probabilistic algorithms, tools for model evaluation, and visualizations to support data exploration. 
+to analyze and model complex network data. The package integrates code implementations from several scientific publications, supporting tasks such as community detection, anomaly detection, and synthetic data generation using latent variable models. It is designed to simplify the use of cutting-edge techniques in network analysis by providing a cohesive and user-friendly interface. 
 
 # Statement of need
 
-Network analysis plays a central role in fields such as social sciences, biology, and fraud detection, where understanding relationships between entities is critical. Probabilistic generative models [@contisciani2020community; @safdari2021generative; @contisciani2022community; @safdari2022anomaly; @safdari2022reciprocity] have emerged as powerful tools for discovering hidden patterns in networks, detecting communities, identifying anomalies, and generating realistic synthetic data.  Despite their potential, the practical use of these models remains challenging due to a lack of integration and accessibility. These methods are often implemented in fragmented codebases spread across individual publications, creating barriers for researchers and practitioners who wish to compare models, reproduce results, or apply them to their own data. ProbINet addresses this critical gap by consolidating recent approaches into a single, unified framework. It provides accessible tools for network analysis tasks, allowing users to explore advanced techniques without the overhead of navigating multiple repositories or inconsistent documentation.  By integrating multiple models and workflows, this package promotes reproducibility, simplifies adoption, and enhances usability across disciplines.
-
-# Mathematical background 
-
-The mathematical foundation of our package builds on recent developments in probabilistic generative models for networks. These models assume that observed network structures arise from underlying latent variables and allow for flexible probabilistic modeling of joint distributions between data and latent variables. By relaxing several restrictive assumptions commonly made in earlier models, our framework supports more expressive methods to uncover hidden structures (e.g., communities and anomalies), model uncertainty, and generate realistic synthetic network data.
+Network analysis plays a central role in fields such as social sciences, biology, and fraud 
+detection, where understanding relationships between entities is critical. Probabilistic 
+generative models [@contisciani2020community; @safdari2021generative; @contisciani2022community; 
+@safdari2022anomaly; @safdari2022reciprocity] have emerged as powerful tools for discovering 
+hidden patterns in networks, detecting communities, identifying anomalies, and generating 
+realistic synthetic data.  However, their use is hindered by fragmented implementations, making 
+comparison and reproduction difficult. ProbINet addresses this critical gap by consolidating 
+recent approaches into a single, unified framework, allowing users to explore advanced techniques without the overhead of navigating multiple repositories or inconsistent documentation, boosting  reproducibility and usability across disciplines.
 
 # Main features
-ProbINet offers a versatile and feature-rich framework to perform inference on networks 
[…]
