### 2025-03-17 #7900 @diegoabt
> [@diegoabt](https://github.com/diegoabt) - also, you describe this as a meta-package of previously published work. Could you explain this a bit more? Have the code in these packages been previously reviewed? Are the authors of these packages co-authors on your submission? Thanks.

Yes, @crvernon . The package consists of the tools presented in 5 different publications. These publications were written by, among others, Caterina De Bacco and Martina Contisciani. They are also the co-authors of the paper I submitted to JOSS. The packages weren't reviewed; they were just published to support the evidence presented in the corresponding papers. We have, among other things, adapted the code and given it a new design. The packages can be found in the following repositories:

1. https://github.com/mcontisc/MTCOV
2. https://github.com/mcontisc/CRep/
3. https://github.com/mcontisc/JointCRep
4. https://github.com/hds-safdari/DynCRep
5. https://github.com/hds-safdari/Anomaly_Community_Detection
