### 2025-03-18 eccb512
Reduced length and added citation.
