+++ publication/paper.bib
@@ -1,66 +1,76 @@
 @article{contisciani2020community,
-  title={Community detection with node attributes in multilayer networks},
-  author={Contisciani, Martina and Power, Eleanor A and De Bacco, Caterina},
-  journal={Scientific reports},
-  volume={10},
-  number={1},
-  pages={15736},
-  year={2020},
-  doi = {10.1038/s41598-020-72626-y},
-  publisher={Nature Publishing Group UK London}
+  title     = {Community detection with node attributes in multilayer networks},
+  author    = {Contisciani, Martina and Power, Eleanor A and De Bacco, Caterina},
+  journal   = {Scientific reports},
+  volume    = {10},
+  number    = {1},
+  pages     = {15736},
+  year      = {2020},
+  doi       = {10.1038/s41598-020-72626-y},
+  publisher = {Nature Publishing Group UK London}
 }
 
 @article{safdari2021generative,
-  title={Generative model for reciprocity and community detection in networks},
-  author={Safdari, Hadiseh and Contisciani, Martina and De Bacco, Caterina},
-  journal={Physical Review Research},
-  volume={3},
-  number={2},
-  pages={023209},
-  year={2021},
-  doi = {10.1103/PhysRevResearch.3.023209},
-  publisher={APS}
+  title     = {Generative model for reciprocity and community detection in networks},
+  author    = {Safdari, Hadiseh and Contisciani, Martina and De Bacco, Caterina},
+  journal   = {Physical Review Research},
+  volume    = {3},
+  number    = {2},
+  pages     = {023209},
+  year      = {2021},
+  doi       = {10.1103/PhysRevResearch.3.023209},
+  publisher = {APS}
 }
 
 @article{contisciani2022community,
-  title={Community detection and reciprocity in networks by jointly modelling pairs of edges},
-  author={Contisciani, Martina and Safdari, Hadiseh and De Bacco, Caterina},
-  journal={Journal of Complex Networks},
-  volume={10},
-  number={4},
-  pages={cnac034},
-  year={2022},
-  doi={10.1093/comnet/cnac034},
-  publisher={Oxford University Press}
+  title     = {Community detection and reciprocity in networks by jointly modelling pairs of edges},
+  author    = {Contisciani, Martina and Safdari, Hadiseh and De Bacco, Caterina},
+  journal   = {Journal of Complex Networks},
+  volume    = {10},
+  number    = {4},
+  pages     = {cnac034},
+  year      = {2022},
+  doi       = {10.1093/comnet/cnac034},
+  publisher = {Oxford University Press}
 }
 
 @article{safdari2022reciprocity,
-  title={Reciprocity, community detection, and link prediction in dynamic networks},
-  author={Safdari, Hadiseh and Contisciani, Martina and De Bacco, Caterina},
-  journal={Journal of Physics: Complexity},
-  volume={3},
-  number={1},
-  pages={015010},
-  year={2022},
-  doi = {10.1088/2632-072X/ac52e6},
-  publisher={IOP Publishing}
+  title     = {Reciprocity, community detection, and link prediction in dynamic networks},
+  author    = {Safdari, Hadiseh and Contisciani, Martina and De Bacco, Caterina},
+  journal   = {Journal of Physics: Complexity},
+  volume    = {3},
+  number    = {1},
+  pages     = {015010},
+  year      = {2022},
+  doi       = {10.1088/2632-072X/ac52e6},
+  publisher = {IOP Publishing}
 }
 
 @article{safdari2022anomaly,
-  title={Anomaly detection and community detection in networks},
-  author={Safdari, Hadiseh and De Bacco, Caterina},
[…]
