+++ publication/paper.md
@@ -41,20 +41,22 @@ to analyze and model complex network data. The package integrates code implement
 
 # Statement of need
 
-Network analysis plays a central role in fields such as social sciences, biology, and fraud 
-detection, where understanding relationships between entities is critical. Probabilistic 
-generative models [@contisciani2020community; @safdari2021generative; @contisciani2022community; 
-@safdari2022anomaly; @safdari2022reciprocity] have emerged as powerful tools for discovering 
-hidden patterns in networks, detecting communities, identifying anomalies, and generating 
-realistic synthetic data.  However, their use is hindered by fragmented implementations, making 
-comparison and reproduction difficult. ProbINet addresses this critical gap by consolidating 
-recent approaches into a single, unified framework, allowing users to explore advanced techniques without the overhead of navigating multiple repositories or inconsistent documentation, boosting  reproducibility and usability across disciplines.
+Network analysis plays a central role in fields such as social sciences, biology, and fraud
+detection, where understanding relationships between entities is critical. Probabilistic
+generative models [@contisciani2020community; @safdari2021generative; @contisciani2022community;
+@safdari2022anomaly; @safdari2022reciprocity] have emerged as powerful tools for discovering
+hidden patterns in networks, detecting communities, identifying anomalies, and generating
+realistic synthetic data.  However, their use is hindered by fragmented implementations, making
+comparisons difficult. ProbINet addresses this critical gap by consolidating
+recent approaches into a single, unified framework, allowing users to explore advanced techniques
+without the overhead of navigating multiple repositories or inconsistent documentation,
+boosting reproducibility and usability across disciplines.
 
 # Main features
 
 ProbINet offers a versatile and feature-rich framework to perform inference on networks using probabilistic generative models.  Key features include:
 
-- **Diverse Network Models**: The package integrates generative models for various network types 
+- **Diverse Network Models**: Integration of generative models for various network types
   and goals:
 
 | **Algorithm's Name**&nbsp; | **Description**                                                                                                         | **Network Properties**                                |
@@ -69,25 +71,21 @@ ProbINet offers a versatile and feature-rich framework to perform inference on n
 |                            |                                                                                                                         |                                                       |
 | **ACD**                    | Identifies anomalous edges and node community memberships in weighted networks [@safdari2022anomaly].                                     | Directed, Weighted, Communities, Anomalies            |
 
-- **Synthetic Network Generation**: ProbINet enables users to generate synthetic networks that closely resemble the real ones for further analyses, such as testing  hypotheses.
+- **Synthetic Network Generation**: Ability to generate synthetic networks that closely resemble real ones for further analyses (e.g., testing hypotheses).
 
-- **Simplified Parameter Selection**: ProbINet includes a cross-validation  module to optimize key parameters, providing performance results in a clear dataframe.
+- **Simplified Parameter Selection**: A cross-validation module to optimize key parameters, providing performance results in a clear dataframe.
 
-- **Rich Set of Metrics for Analysis**:  ProbINet includes metrics like F1 scores, Jaccard index, and advanced metrics for link and covariate prediction performance.
+- **Rich Set of Metrics for Analysis**:  Advanced metrics (e.g., F1 scores, Jaccard index) for link and covariate prediction performance.
 
-- **Powerful Visualization Tools**:  ProbINet includes functions to plot community memberships, and performance metrics.
+- **Powerful Visualization Tools**: Functions for plotting community memberships and performance metrics.
 
[…]
