### 2025-03-18 #7900 @diegoabt
> > > 👋 [@diegoabt](https://github.com/diegoabt) - please fix the ORCID error above in your paper and reduce the word count of your paper to <= 1000 words while I am getting you set up with a topic editor. Thank you!
> >
> >
> > Sure! thank you for pointing this out [@crvernon](https://github.com/crvernon)
>
> [@crvernon](https://github.com/crvernon) , I'm working on this on another branch, and once it's ready I will merge into `master`. I will let you know here once it is ready.

@crvernon , the submission has been updated, and has now less than a 1000 words 😃 .
