### 2025-04-15 8cdc5fd
Add logo as pdf
