+++ docs/source/conf.py
@@ -48,7 +48,7 @@ extensions = [
     "myst_nb",  # add Jupyter notebooks
     "sphinx.ext.napoleon",  # 'sphinxcontrib.napoleon' is deprecated  # Google style docs
     "sphinxcontrib.bibtex",
-    "sphinx_copybutton" # adds a copy button to code blocks
+    "sphinx_copybutton",  # adds a copy button to code blocks
 ]
 autodoc_default_options = {
     "members": True,  # when set to True, Sphinx will automatically document all members
@@ -74,7 +74,7 @@ pygments_style = "sphinx"
 nb_execution_timeout = 1000  # This is the timeout for executing a notebook cell
 
 # -- Options for  sphinxcontrib extension ---------------------------------------
-bibtex_bibfiles = ['references.bib']
+bibtex_bibfiles = ["references.bib"]
 
 # -- Options for HTML evaluation -------------------------------------------------
 
@@ -89,4 +89,4 @@ html_theme = "sphinx_book_theme"
 html_static_path = ["_static"]
 
 # Add the path to the logo PDF file
-html_logo = "_static/logo-probinet.svg"
+html_logo = "_static/logo-probinet.pdf"
