### 2025-04-15 44379ee
Add logo as jpg; use it in docs
