+++ _sources/generated/probinet.model_selection.jointcrep_cross_validation.rst
@@ -1,0 +1,12 @@
+﻿probinet.model\_selection.jointcrep\_cross\_validation
+======================================================
+
+.. automodule:: probinet.model_selection.jointcrep_cross_validation
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      JointCRepCrossValidation
+   
