+++ _sources/generated/probinet.model_selection.crep_cross_validation.rst
@@ -1,0 +1,12 @@
+﻿probinet.model\_selection.crep\_cross\_validation
+=================================================
+
+.. automodule:: probinet.model_selection.crep_cross_validation
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      CRepCrossValidation
+   
