+++ _sources/generated/probinet.models.mtcov.rst
@@ -1,0 +1,12 @@
+﻿probinet.models.mtcov
+=====================
+
+.. automodule:: probinet.models.mtcov
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      MTCOV
+   
