+++ _sources/generated/probinet.models.crep.rst
@@ -1,0 +1,12 @@
+﻿probinet.models.crep
+====================
+
+.. automodule:: probinet.models.crep
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      CRep
+   
