+++ _sources/generated/probinet.model_selection.acd_cross_validation.rst
@@ -1,0 +1,12 @@
+﻿probinet.model\_selection.acd\_cross\_validation
+================================================
+
+.. automodule:: probinet.model_selection.acd_cross_validation
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      ACDCrossValidation
+   
