+++ _sources/generated/probinet.synthetic.dynamic.rst
@@ -1,0 +1,19 @@
+﻿probinet.synthetic.dynamic
+==========================
+
+.. automodule:: probinet.synthetic.dynamic
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      eq_c
+      membership_vectors
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      SyntheticDynCRep
+   
