+++ _sources/generated/probinet.main.rst
@@ -1,0 +1,13 @@
+﻿probinet.main
+=============
+
+.. automodule:: probinet.main
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      main
+      parse_args
+   
