+++ _sources/generated/probinet.models.base.rst
@@ -1,0 +1,14 @@
+﻿probinet.models.base
+====================
+
+.. automodule:: probinet.models.base
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      ModelBase
+      ModelBaseParameters
+      ModelUpdateMixin
+   
