### 2025-05-26 #7900 @anastassiavybornova
Hi @diegoabt ! I will be editing this submission. My sincere apologies for the delay in getting started. I will have a quick look (filling out the checklist below) and will then proceed to looking for reviewers. I can see you already suggested dataspider, Jpickard1, and ClaudMor as reviewers - please feel free to add a couple more names (potentially from [here](https://reviewers.joss.theoj.org/lookup))

Editor checklist (@anastassiavybornova)

**Software**
- [ ] ∃ software description
- [ ] ∃ OSI-approved license
- [ ] ∃ installation instructions
- [ ] ∃ documentation
- [ ] ∃ tests

- [ ] Paper compiles
- [ ] Paper length 250-1000 words

**Application**
- [ ] Substantial scholarly effort
