### 2025-07-13 #7900 @anastassiavybornova
@giorgospanay @sandorjuhasz @szhorvat @mikk-c @sdesojo would you be available to review this submission to JOSS? (if it's your first time reviewing for joss - here is a great way to start: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html )
