### 2025-07-24 #8638 @editorialbot
[REVIEW]: ProbINet: Bridging Usability Gaps in Probabilistic Network Analysis
**Submitting author:** @diegoabt (<a href="http://orcid.org/0000-0003-2994-0138">Diego Baptista</a>)
**Repository:** https://github.com/MPI-IS/probinet
**Branch with paper.md** (empty if default branch):
**Version:** 1.1
**Archive:** 10.17617/3.AYGXPP
