### 2025-07-24 #8638 @anastassiavybornova
@giorgospanay @sandorjuhasz

Dear Giorgos & Sandor,
huge thanks for agreeing to review this submission to JOSS!

Instructions on next steps for you as reviewers (starting by generating the checklist) are already posted by the editorialbot in the comments above, and can also be found here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html# . ping me (@anastassiavybornova ) if you need help or clarifications with the reviewing process; and ping the submitting author, Diego,  @diegoabt if you have questions and/or once you've added your review here.
