### 2025-08-12 MPI-IS/probinet#12 @giorgospanay
[REVIEW] Errors on unit testing
Running the unit tests as described in the documentation, I get 3 failed tests and 4 skipped -- please clarify if this is intended / needs to be fixed, or if any special instructions are needed to execute these.

Stack trace follows (warnings about non-convergence / parameters not saved omitted, I can provide the full trace if necessary):

```
FAIL: test_crep_cross_validation (tests.test_model_selection.TestCrossValidationModels)
Traceback (most recent call last):
  File "/Users/punksnubben/reviews/2025_joss/probinet/tests/test_model_selection.py", line 73, in test_crep_cross_validation
    self.run_cv_and_check_results("CRep")
  File "/Users/punksnubben/reviews/2025_joss/probinet/tests/test_model_selection.py", line 61, in run_cv_and_check_results
    self.assertAlmostEqual(generated_value, ground_truth_value, places=5)
AssertionError: 0.52052381950117 != 0.5205173949559344 within 5 places (6.4245452355837784e-06 difference)

FAIL: test_dyncrep_cross_validation (tests.test_model_selection.TestCrossValidationModels)
Traceback (most recent call last):
  File "/Users/punksnubben/reviews/2025_joss/probinet/tests/test_model_selection.py", line 64, in test_dyncrep_cross_validation
    self.run_cv_and_check_results("DynCRep")
  File "/Users/punksnubben/reviews/2025_joss/probinet/tests/test_model_selection.py", line 61, in run_cv_and_check_results
    self.assertAlmostEqual(generated_value, ground_truth_value, places=5)
AssertionError: 0.7280598583405772 != 0.7280397271883402 within 5 places (2.0131152236979588e-05 difference)

FAIL: test_running_algorithm_with_rng_seed_0 (tests.test_validation_JointCRep.JointCRepTestCase)
Traceback (most recent call last):
  File "/Users/punksnubben/reviews/2025_joss/probinet/tests/test_validation_JointCRep.py", line 137, in test_running_algorithm_with_rng_seed_0
    self.assertTrue(
AssertionError: False is not true : The 'w' values should be equal for the same random seed

Ran 139 tests in 1138.178s

FAILED (failures=3, skipped=4)
```

openjournals/joss-reviews#8638
