### 2025-08-12 #8638 @giorgospanay
Hello @diegoabt, thank you for the opportunity to review this software. It appears quite useful and well developed, and has met most of the checklist items on my end. I just have a few minor points requiring clarification:

_Functionality:_ The tutorials run successfully on my device. However, when executing the unit tests, I encountered 3 failures and 4 skipped tests. I am unsure whether this behavior is expected -- please refer to the trace linked in the open issue for details.
_State of the Field:_ The paper does not appear to mention whether other relevant software for probabilistic networks exists (or explicitly state that none does). Could you please clarify this point?
