### 2025-09-20 MPI-IS/probinet#12 @diegoabt
Hi @giorgospanay , thank you for reporting this issue. I've created a new branch to solve these problems (not visible yet). I solved the first two by increasing the tolerance; 1e-5 is already a very tight precision, 1e-4 should do the job. Regarding the error about the parameter `w`, I'm not sure how this could be happening, given that we test our package in multiple CIs. Could you please tell me which python version you are using, and which OS you have? Thank you.
