### 2025-09-21 MPI-IS/probinet#12 @giorgospanay
Hi @diegoabt , sounds good! I ran the tests on macOS 15.6.1, Python 3.10.5 (using pyenv for version switching). I just ran the tests again and reproduced the issues. Posting also the output in case any of the previous errors/warnings are relevant.

```
..ss....../Users/punksnubben/reviews/2025_joss/probinet/venv/lib/python3.10/site-packages/sklearn/metrics/_ranking.py:1192: UndefinedMetricWarning: No negative samples in y_true, false positive value should be meaningless
  warnings.warn(
...........ERROR:root:There are negative weights.
..WARNING:root:Solution failed to converge in 500 EM steps!
WARNING:root:Parameters won't be saved for this realization! If you have provided a number of realizations equal to one, please increase it.
.WARNING:root:Solution failed to converge in 500 EM steps!
WARNING:root:Parameters won't be saved for this realization! If you have provided a number of realizations equal to one, please increase it.
.........................WARNING:root:Solution failed to converge in 500 EM steps!
WARNING:root:Parameters won't be saved for this realization! If you have provided a number of realizations equal to one, please increase it.
WARNING:root:Solution failed to converge in 500 EM steps!
WARNING:root:Parameters won't be saved for this realization! If you have provided a number of realizations equal to one, please increase it.
.FF.WARNING:root:Solution failed to converge in 500 EM steps!
WARNING:root:Parameters won't be saved for this realization! If you have provided a number of realizations equal to one, please increase it.
WARNING:root:Solution failed to converge in 500 EM steps!
WARNING:root:Parameters won't be saved for this realization! If you have provided a number of realizations equal to one, please increase it.
WARNING:root:Solution failed to converge in 500 EM steps!
WARNING:root:Parameters won't be saved for this realization! If you have provided a number of realizations equal to one, please increase it.
WARNING:root:Solution failed to converge in 500 EM steps!
WARNING:root:Parameters won't be saved for this realization! If you have provided a number of realizations equal to one, please increase it.
............................ERROR:root:The Normalization parameter can be either 0 or 1! It is used as an indicator for generating the membership matrices u and v from a Dirichlet or a Gamma distribution, respectively. It is used when there is overlapping.
.ERROR:root:The correlation parameter corr has to be in [0, 1]!
.ERROR:root:The reciprocity coefficient eta has to be in [0, 1)!
.ERROR:root:The overlapping parameter has to be in [0, 1]!
.ERROR:root:The structure of the affinity matrix w can be either assortative or disassortative!
.......s../Users/punksnubben/reviews/2025_joss/probinet/probinet/synthetic/multilayer.py:204: UserWarning: avg_degree parameter was not set. Defaulting to avg_degree=15
  warnings.warn(msg)
/Users/punksnubben/reviews/2025_joss/probinet/probinet/synthetic/multilayer.py:213: UserWarning: directed parameter was not set. Defaulting to directed=False
  warnings.warn(msg)
/Users/punksnubben/reviews/2025_joss/probinet/probinet/synthetic/multilayer.py:221: UserWarning: is_sparse parameter was not set. Defaulting to is_sparse=True
  warnings.warn(msg)
/Users/punksnubben/reviews/2025_joss/probinet/probinet/synthetic/multilayer.py:229: UserWarning: label parameter was not set. Defaulting to label=_N_L_K_avgdegree
  warnings.warn(msg)
/Users/punksnubben/reviews/2025_joss/probinet/probinet/synthetic/multilayer.py:252: UserWarning: perc_overlapping parameter was not set. Defaulting to perc_overlapping=0.4
  warnings.warn(msg)
/Users/punksnubben/reviews/2025_joss/probinet/probinet/synthetic/multilayer.py:268: UserWarning: correlation_u_v parameter for overlapping communities was not set. Defaulting to corr=0.0
  warnings.warn(msg)
/Users/punksnubben/reviews/2025_joss/probinet/probinet/synthetic/multilayer.py:276: UserWarning: alpha parameter of Dirichlet distribution was not set. Defaulting to alpha=[0.1, 0.1, 0.1]
  warnings.warn(msg)
.......ERROR:root:Cannot cast 12.34 to integer.
ERROR:root:Cannot cast abc to integer.
..............WARNING:root:Solution failed to converge in 500 EM steps!
WARNING:root:Parameters won't be saved for this realization! If you have provided a number of realizations equal to one, please increase it.
.......WARNING:root:Solution failed to converge in 1000 EM steps!
WARNING:root:Parameters won't be saved for this realization! If you have provided a number of realizations equal to one, please increase it.
...s... ....F.WARNING:root:Solution failed to converge in 1000 EM steps!
WARNING:root:Parameters won't be saved for this realization! If you have provided a number of realizations equal to one, please increase it.
...
FAIL: test_crep_cross_validation (tests.test_model_selection.TestCrossValidationModels)
Traceback (most recent call last):
  File "/Users/punksnubben/reviews/2025_joss/probinet/tests/test_model_selection.py", line 73, in test_crep_cross_validation
    self.run_cv_and_check_results("CRep")
  File "/Users/punksnubben/reviews/2025_joss/probinet/tests/test_model_selection.py", line 61, in run_cv_and_check_results
    self.assertAlmostEqual(generated_value, ground_truth_value, places=5)
AssertionError: 0.52052381950117 != 0.5205173949559344 within 5 places (6.4245452355837784e-06 difference)

FAIL: test_dyncrep_cross_validation (tests.test_model_selection.TestCrossValidationModels)
Traceback (most recent call last):
  File "/Users/punksnubben/reviews/2025_joss/probinet/tests/test_model_selection.py", line 64, in test_dyncrep_cross_validation
    self.run_cv_and_check_results("DynCRep")
  File "/Users/punksnubben/reviews/2025_joss/probinet/tests/test_model_selection.py", line 61, in run_cv_and_check_results
    self.assertAlmostEqual(generated_value, ground_truth_value, places=5)
AssertionError: 0.7280598583405772 != 0.7280397271883402 within 5 places (2.0131152236979588e-05 difference)

FAIL: test_running_algorithm_with_rng_seed_0 (tests.test_validation_JointCRep.JointCRepTestCase)
Traceback (most recent call last):
  File "/Users/punksnubben/reviews/2025_joss/probinet/tests/test_validation_JointCRep.py", line 137, in test_running_algorithm_with_rng_seed_0
    self.assertTrue(
AssertionError: False is not true : The 'w' values should be equal for the same random seed

Ran 139 tests in 1020.035s

FAILED (failures=3, skipped=4)
```
