### 2025-09-21 #8638 @anastassiavybornova
hi @diegoabt thanks for checking back. let me try to ping reviewer nr 2 one more time.

@sandorjuhasz would you still be available for reviewing this joss submission?

huge thanks in advance!
