### 2025-09-22 982be89
Work on review
Add extra references

Relax checks for w in JointCRep validation test

Remove skipped tests

Edit paper.md
Reduce length of paper.md
