+++ publication/paper.bib
@@ -74,3 +74,36 @@
   version = {1.16.1},
   year    = {2023}
 }
+
+@article{peixoto_graph-tool_2014,
+  title        = {The graph-tool python library},
+  author       = {Peixoto, Tiago P.},
+  journal      = {figshare},
+  year         = {2014},
+  doi          = {10.6084/m9.figshare.1164194},
+  url          = {http://figshare.com/articles/graph_tool/1164194},
+  keywords     = {graph, network, tools}
+}
+
+@article{rossetti_cdlib_2019,
+  title        = {CDlib: a Python Library to Extract, Compare and Evaluate Communities from Complex Networks},
+  author       = {Rossetti, Giulio and Milli, Letizia and Cazabet, Rémy},
+  journal      = {Applied Network Science},
+  year         = {2019},
+  volume       = {4},
+  number       = {1},
+  pages        = {52},
+  doi          = {10.1007/s41109-019-0165-9},
+  url          = {https://doi.org/10.1007/s41109-019-0165-9}
+}
+
+@article{ankan_pgmpy_2024,
+  author    = {Ankur Ankan and Johannes Textor},
+  title     = {pgmpy: A Python Toolkit for Bayesian Networks},
+  journal   = {Journal of Machine Learning Research},
+  year      = {2024},
+  volume    = {25},
+  number    = {265},
+  pages     = {1--8},
+  url       = {http://jmlr.org/papers/v25/23-0487.html}
+}
