### 2025-09-22 MPI-IS/probinet#12 @diegoabt
Great @giorgospanay ! thank you for sending the full log. I fixed the errors and verified the changes on our CI. The test was failing because it required two arrays to be exactly equal. Since `default_rng` was upgraded, its random outputs changed. Replacing the strict equality check with `np.allclose` at high precision resolves the issue while preserving result reliability.
