### 2025-09-22 MPI-IS/probinet#12 @giorgospanay
Running the tests again, the last error is resolved, but I still get the assertion failure on dyncrep -- maybe adjusting the precision to 3 decimals (or perhaps a delta of 10^-4?) should do it. Log follows:

```
FAIL: test_dyncrep_cross_validation (tests.test_model_selection.TestCrossValidationModels)
Traceback (most recent call last):
  File "/Users/punksnubben/reviews/2025_joss/probinet/tests/test_model_selection.py", line 64, in test_dyncrep_cross_validation
    self.run_cv_and_check_results("DynCRep")
  File "/Users/punksnubben/reviews/2025_joss/probinet/tests/test_model_selection.py", line 61, in run_cv_and_check_results
    self.assertAlmostEqual(generated_value, ground_truth_value, places=4)
AssertionError: 0.8321948137487244 != 0.8322667379202814 within 4 places (7.192417155699182e-05 difference)

Ran 135 tests in 1032.545s

FAILED (failures=1)
```
