### 2025-09-22 22d8154
Fix tests for model validation
