### 2025-11-02 #8638 @diegoabt
Hi @sandorjuhasz ,

Thank you so much for you review and your nice comments. We are all happy to know you find the package useful and well documented.

Regarding the minor comments:
1. Sounds good. I think we can sort them based on how easy it is to use them.
2. I see how adding more information about other algorithms might help the users to better understand the contributions made by our package. However, we have already addressed a similar comment from a previous reviewer and, within the space constraints of the JOSS format, we compared probinet to three related algorithms: graph-tool, CDlib, and pgmpy. Expanding this section further would likely require removing or condensing other essential parts of the paper.

That said, we would be happy to consider adding a concise statement that briefly situates probinet in relation to other commonly used libraries if you have specific aspects or relationships you think should be emphasized.
