### 2025-11-07 dbaf7f9
Implement suggestions made by reviewer
