+++ publication/paper.md
@@ -12,22 +12,24 @@ authors:
   - name: Diego Baptista
     orcid: 0000-0003-2994-0138
     #equal-contrib: true
-    affiliation: 1 
+    affiliation: 1,2
   - name: Martina Contisciani
     #corresponding: true # (This is how to denote the corresponding author)
-    affiliation: 2
+    affiliation: 3
   - name: Caterina De Bacco
     #equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
-    affiliation: 3
+    affiliation: 4
   - name: Jean-Claude Passy
     affiliation: 1
 affiliations:
   - name: Max Planck Institute for Intelligent Systems, Tübingen, Germany.
     index: 1
-  - name: Central European University, Vienna, Austria.
+  - name: Graz University of Technology, Graz, Austria
     index: 2
-  - name: Delft University of Technology, Delft, Netherlands.
+  - name: Central European University, Vienna, Austria.
     index: 3
+  - name: Delft University of Technology, Delft, Netherlands.
+    index: 4
 date: 22 January 2025
 bibliography: paper.bib
 
@@ -42,7 +44,9 @@ to analyze and model complex network data. The package integrates code implement
 # Statement of need
 
 Network analysis is central to social sciences, biology, and fraud detection, where 
-understanding relationships is essential. Probabilistic generative models [@contisciani2020community; @safdari2021generative; @contisciani2022community; @safdari2022anomaly; @safdari2022reciprocity] reveal hidden patterns, detect communities, identify anomalies, and generate synthetic data. Their broader use is limited by fragmented implementations that hinder comparisons and reproducibility. 
+understanding relationships is essential. Probabilistic generative models 
+[@safdari2021generative; @contisciani2022community; @safdari2022anomaly; @safdari2022reciprocity; @contisciani2020community
+] reveal hidden patterns, detect communities, identify anomalies, and generate synthetic data. Their broader use is limited by fragmented implementations that hinder comparisons and reproducibility. 
 ProbINet addresses this gap by unifying recent approaches in a single framework, improving accessibility and usability across disciplines. 
 
 ProbINet stands out among network analysis tools. Graph-tool [@peixoto_graph-tool_2014] provides community detection and general graph analysis tools, but it uses a different model family than our mixed-membership framework and does not account for reciprocity.  CDlib [@rossetti_cdlib_2019] offers detection algorithms and evaluation routines, but ProbINet extends this with probabilistic MLE models, optional node attributes, and anomaly detection. pgmpy [@ankan_pgmpy_2024] focuses on Bayesian network structure learning, while ProbINet uncovers latent patterns like communities and reciprocity.
@@ -57,8 +61,6 @@ generative models.  Key features include:
 
 | **Algorithm's Name**&nbsp; | **Description**                                                                                                         | **Network Properties**                                |
 |----------------------------|-------------------------------------------------------------------------------------------------------------------------|-------------------------------------------------------|
-| **MTCOV**                  | Extracts overlapping communities in multilayer networks using topology and node attributes [@contisciani2020community]. | Weighted, Multilayer, Attributes, Communities         |
-|                            |                                                                                                                         |                                                       |
 | **CRep**                   | Models directed networks with communities and reciprocity [@safdari2021generative].                                     | Directed, Weighted, Communities, Reciprocity          |
 |                            |                                                                                                                         |                                                       |
 | **JointCRep**              | Captures community structure and reciprocity with a joint edge distribution [@contisciani2022community].                | Directed, Communities, Reciprocity                    |
@@ -66,6 +68,8 @@ generative models.  Key features include:
[…]
