### 2025-11-07 7fb914d
Update README to match algorithm's order included in paper
