+++ README.md
@@ -27,11 +27,11 @@ The package integrates a wide range of probabilistic  generative models tailored
 
 | **Name**         | **Description**                                                                                       | **Network Properties**                   |
 |-----------------|----------------------------------------------------------------------------------------------------|----------------------------------------|
-| **MTCOV**       | Extracts overlapping communities in multilayer networks using topology and node attributes [CPDB20].       | Weighted, Multilayer, Attributes, Communities      |
 | **CRep**        | Models directed networks with communities and reciprocity [CSDB22].                                                       | Directed, Weighted, Communities, Reciprocity                 |
 | **JointCRep**   | Captures community structure and reciprocity with a joint edge distribution [SCDB21].                   | Directed, Communities, Reciprocity       |
 | **DynCRep**     | Extends CRep for dynamic networks [SCDB22].                                                             | Directed, Weighted, Dynamic, Communities, Reciprocity           |
 | **ACD**         | Identifies anomalous edges and node community memberships in weighted networks [SDB22]. | Directed, Weighted, Communities, Anomalies       |
+| **MTCOV**       | Extracts overlapping communities in multilayer networks using topology and node attributes [CPDB20].       | Weighted, Multilayer, Attributes, Communities      |
 
 For more details on each model, please refer to the [documentation](https://mpi-is.github.io/probinet/). References to the original papers can be found [here](https://mpi-is.github.io/probinet/references.html).
 
