### 2025-11-17 #8638 @anastassiavybornova
@diegoabt

looking great! some last steps for you to tick off:

**paper editing**

- [ ] in the statement of need, first sentence, please include "Network analysis is central to **disciplines such as** social sciences, ..." (to indicate the enumeration is non-exhaustive)
- [ ] in the "Main features" bullet point list, in the first item, the "integration ... for various network types and goals:" is followed by a table which is currently part of the bulletpoint list -- please simplify by providing the table of network generation algorithms separately (as table 1), and then referring to it in the "main features" bulletpoint list by referencing the table.

**archiving**

- [ ] make a new tagged release of the software;
- [ ] archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or other)
- [ ] post the version number and archive DOI here below
