### 2025-11-17 7d4e2d0
Fix position of table describing algorithms
