### 2025-12-23 #8638 @anastassiavybornova
@diegoabt just to let you know regarding timelines, i'll be available on dec 29/30 and will be taking off for several weeks after that, ideally you could update the release and metada as described below by dec 30 at the latest to get this out without further delays -- otherwise no worries and we will do so by end of january '26. best wishes
