+++ _sources/generated/probinet.input.preprocessing.rst
@@ -1,0 +1,15 @@
+﻿probinet.input.preprocessing
+============================
+
+.. automodule:: probinet.input.preprocessing
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      create_adjacency_tensor_from_graph_list
+      create_sparse_adjacency_tensor_from_graph_list
+      preprocess_adjacency_tensor
+      preprocess_data_matrix
+   
