+++ _sources/generated/probinet.synthetic.base.rst
@@ -1,0 +1,21 @@
+﻿probinet.synthetic.base
+=======================
+
+.. automodule:: probinet.synthetic.base
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      affinity_matrix
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      BaseSyntheticNetwork
+      GraphProcessingMixin
+      StandardMMSBM
+      Structure
+   
