+++ _sources/generated/probinet.model_selection.main.rst
@@ -1,0 +1,12 @@
+﻿probinet.model\_selection.main
+==============================
+
+.. automodule:: probinet.model_selection.main
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      cross_validation
+   
