+++ _sources/generated/probinet.models.dyncrep.rst
@@ -1,0 +1,12 @@
+﻿probinet.models.dyncrep
+=======================
+
+.. automodule:: probinet.models.dyncrep
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      DynCRep
+   
