+++ _sources/generated/probinet.model_selection.parameter_search.rst
@@ -1,0 +1,12 @@
+﻿probinet.model\_selection.parameter\_search
+===========================================
+
+.. automodule:: probinet.model_selection.parameter_search
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      define_grid
+   
