+++ _sources/generated/probinet.synthetic.reciprocity.rst
@@ -1,0 +1,13 @@
+﻿probinet.synthetic.reciprocity
+==============================
+
+.. automodule:: probinet.synthetic.reciprocity
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      GM_reciprocity
+      ReciprocityMMSBM_joints
+   
