+++ _sources/generated/probinet.input.loader.rst
@@ -1,0 +1,17 @@
+﻿probinet.input.loader
+=====================
+
+.. automodule:: probinet.input.loader
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      build_adjacency_and_design_from_file
+      build_adjacency_from_file
+      build_adjacency_from_networkx
+      read_and_process_design_matrix
+      read_design_matrix
+      read_graph
+   
