+++ _sources/generated/probinet.model_selection.masking.rst
@@ -1,0 +1,17 @@
+﻿probinet.model\_selection.masking
+=================================
+
+.. automodule:: probinet.model_selection.masking
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      extract_mask_kfold
+      extract_masks
+      shuffle_indices
+      shuffle_indicesG
+      shuffle_indicesX
+      shuffle_indices_all_matrix
+   
