### 2026-06-05 138ff31
Document return values and exceptions for utils primitives.
Add Returns and Raises sections to docstrings for leaf functions in
utils.tools and utils.matrix_operations to clarify outputs and failure
modes without changing runtime behavior.

Co-authored-by: Cursor <cursoragent@cursor.com>
