### 2026-08-13 4b8cb0c
Update manuscript for JOSS submission
