+++ publication/paper.bib
@@ -58,11 +58,11 @@
   publisher = {Springer}
 }
 
-@article{konect:coleman,
-  title   = {Introduction to Mathematical Sociology},
-  author  = {Coleman, James Samuel},
-  journal = {London Free Press Glencoe},
-  year    = {1964}
+@book{konect:coleman,
+  title     = {Introduction to mathematical sociology},
+  author    = {Coleman, James Samuel},
+  publisher = {Free Press of Glencoe},
+  year      = {1964}
 }
 
 @software{Peter_hyperfine_2023,
@@ -76,7 +76,7 @@
 }
 
 @article{peixoto_graph-tool_2014,
-  title        = {The graph-tool python library},
+  title        = {The graph-tool {Python} library},
   author       = {Peixoto, Tiago P.},
   journal      = {figshare},
   year         = {2014},
@@ -86,7 +86,7 @@
 }
 
 @article{rossetti_cdlib_2019,
-  title        = {CDlib: a Python Library to Extract, Compare and Evaluate Communities from Complex Networks},
+  title        = {CDlib: a {Python} library to Extract, Compare and Evaluate Communities from Complex Networks},
   author       = {Rossetti, Giulio and Milli, Letizia and Cazabet, Rémy},
   journal      = {Applied Network Science},
   year         = {2019},
@@ -99,11 +99,43 @@
 
 @article{ankan_pgmpy_2024,
   author    = {Ankur Ankan and Johannes Textor},
-  title     = {pgmpy: A Python Toolkit for Bayesian Networks},
+  title     = {{pgmpy}: A {Python} toolkit for {Bayesian} networks},
   journal   = {Journal of Machine Learning Research},
   year      = {2024},
   volume    = {25},
   number    = {265},
   pages     = {1--8},
   url       = {http://jmlr.org/papers/v25/23-0487.html}
+}
+
+@article{mcfarland2014network,
+  title     = {Network ecology and adolescent social structure},
+  author    = {McFarland, Daniel A. and Moody, James and Diehl, David and Smith, Jeffrey A. and Thomas, Reuben J.},
+  journal   = {American Sociological Review},
+  volume    = {79},
+  number    = {6},
+  pages     = {1088--1121},
+  year      = {2014},
+  doi       = {10.1177/0003122414554001}
+}
+
+@article{vaquera2008friendship,
+  title     = {Do you like me as much as {I} like you? {Friendship} reciprocity and its effects on school outcomes among adolescents},
+  author    = {Vaquera, Elizabeth and Kao, Grace},
+  journal   = {Social Science Research},
+  volume    = {37},
+  number    = {1},
+  pages     = {55--72},
+  year      = {2008},
+  doi       = {10.1016/j.ssresearch.2006.11.002}
+}
+
+@article{block2015reciprocity,
+  title     = {Reciprocity, transitivity, and the mysterious three-cycle},
+  author    = {Block, Per},
+  journal   = {Social Networks},
+  volume    = {40},
+  pages     = {163--173},
+  year      = {2015},
+  doi       = {10.1016/j.socnet.2014.10.005}
 }
